+++ com.pilosa.client/src/main/java/com/pilosa/client/BoolResult.java
@@ -70,6 +70,16 @@ public class BoolResult implements QueryResult {
     }
 
     @Override
+    public List<GroupCount> getGroupCounts() {
+        return GroupCountsResult.defaultItems();
+    }
+
+    @Override
+    public RowIdentifiersResult getRowIdentifiers() {
+        return RowIdentifiersResult.defaultResult();
+    }
+
+    @Override
     public boolean equals(Object obj) {
         if (obj == this) {
             return true;