+++ com.pilosa.client/src/main/java/com/pilosa/client/RowIdentifiersResult.java
@@ -37,12 +37,14 @@ package com.pilosa.client;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 
+import java.util.ArrayList;
 import java.util.List;
 
-public class ValueCountResult implements QueryResult {
+public final class RowIdentifiersResult implements QueryResult {
+
     @Override
     public int getType() {
-        return QueryResultType.VAL_COUNT;
+        return QueryResultType.GROUP_COUNTS;
     }
 
     @Override
@@ -57,12 +59,12 @@ public class ValueCountResult implements QueryResult {
 
     @Override
     public long getCount() {
-        return this.count;
+        return 0;
     }
 
     @Override
     public long getValue() {
-        return this.value;
+        return 0;
     }
 
     @Override
@@ -77,7 +79,7 @@ public class ValueCountResult implements QueryResult {
 
     @Override
     public RowIdentifiersResult getRowIdentifiers() {
-        return RowIdentifiersResult.defaultResult();
+        return this;
     }
 
     @Override
@@ -85,36 +87,45 @@ public class ValueCountResult implements QueryResult {
         if (obj == this) {
             return true;
         }
-        if (!(obj instanceof ValueCountResult)) {
+        if (!(obj instanceof RowIdentifiersResult)) {
             return false;
         }
-        ValueCountResult rhs = (ValueCountResult) obj;
+        RowIdentifiersResult rhs = (RowIdentifiersResult) obj;
         return new EqualsBuilder()
-                .append(this.value, rhs.value)
-                .append(this.count, rhs.count)
+                .append(this.rowIDs, rhs.rowIDs)
+                .append(this.rowKeys, rhs.rowKeys)
                 .isEquals();
     }
 
     @Override
     public int hashCode() {
         return new HashCodeBuilder(31, 47)
-                .append(this.value)
-                .append(this.count)
+                .append(this.rowIDs)
+                .append(this.rowKeys)
                 .toHashCode();
     }
 
-    static ValueCountResult create(long sum, long count) {
-        ValueCountResult result = new ValueCountResult();
-        result.value = sum;
-        result.count = count;
-        return result;
+    static RowIdentifiersResult defaultResult() {
+        return defaultResult;
     }
 
-    static ValueCountResult fromInternal(Internal.QueryResult q) {
-        Internal.ValCount obj = q.getValCount();
-        return create(obj.getVal(), obj.getCount());
+    static RowIdentifiersResult fromInternal(Internal.QueryResult q) {
+        Internal.RowIdentifiers rowIdentifiers = q.getRowIdentifiers();
+        return new RowIdentifiersResult(rowIdentifiers.getRowsList(), rowIdentifiers.getKeysList());
     }
 
-    private long value;
-    private long count;
+    static {
+        RowIdentifiersResult result = new RowIdentifiersResult(new ArrayList<Long>(0), new ArrayList<String>(0));
+        defaultResult = result;
+    }
+
+    private RowIdentifiersResult(List<Long> rowIDs, List<String> rowKeys) {
+        this.rowIDs = rowIDs;
+        this.rowKeys = rowKeys;
+    }
+
+    private final static RowIdentifiersResult defaultResult;
+
+    private final List<Long> rowIDs;
+    private final List<String> rowKeys;
 }